requirements.txt
django==3.2.9
djangorestframework==3.14.0
djoser==2.1.0
environs==9.5.0
django-currentuser==0.5.3
